requirements.txt
fastapi==0.116.1
httpx==0.28.1
langchain==0.3.27
langchain_community==0.3.27
langchain_openai==0.3.30
loguru==0.7.2
pyautogen==0.2.0
pydantic==2.11.7
pydantic_settings==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.1
PyYAML==6.0.2
uvicorn==0.35.0
